naked returns

+++ x/rank/types/index.go
@@ -159,7 +159,7 @@ func (i *BaseSearchIndex) Top(page, perPage uint32) ([]RankedCidNumber, uint32, 
 
 // make sure that this link (from-to) is new
 func (i *BaseSearchIndex) handleLink(link graphtypes.CompactLink) {
-	i.extendIndex(uint64(link.From))
+	i.extendIndex(link.From)
 
 	fromIndex := i.links[link.From]
 	// in case unlock signal received we could operate on this index otherwise put link in the end of queue and finish
